+++ src/main/java/io/xlorey/FluxLoader/plugin/Plugin.java
@@ -17,6 +17,24 @@ public class Plugin implements IPlugin {
     }
 
     /**
+     * Plugin activation event
+     * Called when the plugin is enabled through the main game menu
+     */
+    @Override
+    public void onEnable() {
+
+    }
+
+    /**
+     * Plugin deactivation event
+     * Called when the plugin is disabled through the main game menu
+     */
+    @Override
+    public void onDisable() {
+
+    }
+
+    /**
      * Getting information about a plugin
      * @return information about the plugin in PluginInfo format
      */
@@ -28,7 +46,7 @@ public class Plugin implements IPlugin {
      * Setting information about a plugin
      * @param pluginInfo metadata about the plugin in the form of PluginInfo
      */
-    public void setInfo(PluginInfo pluginInfo) {
+    public void setPluginInfo(PluginInfo pluginInfo) {
         this.pluginInfo = pluginInfo;
     }
 }