+++ changelog.md
@@ -1,3 +1,6 @@
+# FluxLoader v0.6.0
+- Added `onEnable` and `onDisable` events for plugins. On the server: `onEnable` is called at startup and `onDisable` - at shutdown. On the client: `onEnable` and `onDisable` are called through the main game menu when enabling/disabling the plugin.
+
 # FluxLoader v0.5.0
 - Added a tool for sending a message from the server to a specific user
 - Added `onPlayerFullyConnected` event
@@ -5,6 +8,7 @@
 - Added packet blocking mechanism to `IncomingPacket.blockPacket()`
 - Added date and time output via `Logger.print()`
 - Fixed error output when calling events
+
 # FluxLoader v0.4.1
 - Added new annotations (CommandName, CommandAccessLevel, CommandChatReturn, CommandExecutionScope) have been created to indicate various aspects of commands.
 - Added AccessLevel and CommandScope enumerations to represent access levels and command scopes, respectively.